Change how errors and rejections are caught in ObservableList promises
Details:

ObservableList promises were re-raising caught errors to make them available
to be caught downstream. However, this resulted in unhandled promise rejection
errors if the user of each ObservableList did not catch these errors.

ObservableList was designed to free developers from worrying about promises,
so we should not force developers to worry about catching promise errors.

This change fixes the above issue by making an intermediate promise available
to users of ObservableList, before chaining a `catch` handler to the promise.
The downstream ObservableList may chain one or more parallel `catch` handlers
to the intermediate promise, but they are not obligated to do so, because the
internals of ObservableList also subsequently chain a `catch` handler to each
of these intermediate promises, which prevents unhandled promise rejections.

The same issue and fixes were made to PaginatedObservableList in this change.